feat: Product Infrastructure 구현
- ProductEntity: JPA 엔티티, Entity ↔ Domain 변환
- ProductJpaRepository: 비관적 락 적용 (findByIdWithLock)
- ProductRepositoryImpl: Domain Repository 구현체
- ProductService, Controller, Response: Domain 모델 사용하도록 수정

+++ server-java/src/main/java/kr/hhplus/be/server/presentation/product/ProductRequest.java
@@ -1,11 +1,9 @@
 package kr.hhplus.be.server.presentation.product;
 
-import java.math.BigDecimal;
-
 public record ProductRequest(
         String name,
         String description,
-        BigDecimal price,
-        int stock
+        Long price,
+        Integer stock
 ) {
 }
